fix request list resizing
The request list could only be shrunk, but not grown, because when the
mouse was over the Django Debug Panel iframe, mouse events would go to
the iframe and not to the parent document.  Fix by setting the iframe's
pointer-events CSS property to "none" while a resize is in progress.